Separate matches method from operator utils
